Add a new class to manage the more complex bit packing of the configuration register. Update everything for the new 17-bit data length (makes a lot of things less clean).
Still thinking about a way to combo the event counter read with
the more efficient 16 bit data length read that is nice.

+++ minervadaq/workers/include/ReadoutWorker.h
@@ -64,8 +64,8 @@ class ReadoutWorker {
   void ResetCurrentChannel();
   bool MoveToNextChannel();
   const EChannels * CurrentChannel() const;
-  unsigned short GetNextDataBlockSize() const;
-  std::tr1::shared_ptr<SequencerReadoutBlock> GetNextDataBlock( unsigned short blockSize ) const;
+  unsigned int GetNextDataBlockSize() const;
+  std::tr1::shared_ptr<SequencerReadoutBlock> GetNextDataBlock( unsigned int blockSize ) const;
 };
 
 #endif
